Add files via upload
Requirements
Design and implement your own, "homemade" IoC container.

Your container will require that components follow a certain convention for dependency injection (you may impose that it only accepts constructor or setter injection). 

A second requirement for your container is that it accepts configuration information given in configuration files. You may choose any format for the syntax of config files (XML files, text files).  The semantics of your config file must be able cover all info that can be found in a component diagram with blackbox components.

 Also, design a small API for clients to retrieve the composition from the container (you may find a source of inspiration in the Spring example).

Special requirements:

Give your container more decision power: the configuration file may contain several matches for a requirement. The container will decide which implementation to use. For example, in the MovieLister-MovieFinder scenario,  suppose there are several MovieFinder Implementations available and registered with the container. The container will pick one to inject into the MovieLister.

Your IoC framework must support several  levels of transitive injection.  It is not only that MovieLister requires a MovieFinder, but, for example, a particular MovieFinder requires a DatabaseAccess,  and the DatabaseAcces further requires a NetworkCommunicator.  Thus, when an application requires to get the MovieLister, your IoC framework will identify that it requires a MovieFinder, and it will find one MovieFinder implementation that will be instantiate and injected. If this Finder has requirements, these must be transitively solved and injected.    

Demonstrate the genericity of your framework by developing 2 different component-based applications on top of it; one of the applications can be the MovieLister+MovieFinder scenario, but extend the scenario such that several levels of transitive requirements should be involved.